net/mlx5: fix interrupt completion queue index wrapping
[ backported from upstream commit 3afdf157fcedba15167ad03b187746bd6c73811f ]

Rxq cq_ci was 16 bits while hardware is expecting to wrap
around 24 bits, this caused interrupt failure after burst of packets.

Fixes: 43e9d9794cde ("net/mlx5: support upstream rdma-core")

Acked-by: Yongseok Koh <yskoh@mellanox.com>

+++ drivers/net/mlx5/mlx5_rxtx.h
@@ -138,9 +138,9 @@ struct mlx5_rxq_data {
 	volatile uint32_t *rq_db;
 	volatile uint32_t *cq_db;
 	uint16_t port_id;
-	uint16_t rq_ci;
-	uint16_t rq_pi;
-	uint16_t cq_ci;
+	uint32_t rq_ci;
+	uint32_t rq_pi;
+	uint32_t cq_ci;
 	uint16_t rq_repl_thresh; /* Threshold for buffer replenishment. */
 	struct mlx5_mr_ctrl mr_ctrl;
 	volatile struct mlx5_wqe_data_seg(*wqes)[];
